requirements.txt
flask_smorest
flask_cors
influxdb_client
requests
flask-talisman
